Updated openstack/openstack Project: openstack/glance  614c0fe9b324b2fb0ed2b16fae8b5488c233df96
Fix levels of Swift configuration documentation

Currently the "Configuring Multiple Swift Accounts/Stores"
section is regarded as a section, but it should be placed in
the "Configuring the Swift Storage Backend" subsection.

In addition, all the glance stores configuration subsections
should be placed under "Configuring Glance Storage Backends"
section, but it is now misplaced by inserted sections.

This moves the sections to valid places.

Change-Id: I24fff19d2a631c8106b996f9a9061f7e50dacece