Updated openstack/openstack Project: openstack/keystone  3eb3dd37bbedb1ad58a9d6934ec60ee4dcd69e4f
Explicitly exclude tests from bandit scan

Now that the bandit linter runs without a bandit.yaml specified,
it should exclude the tests directory similar to how it did before
in the bandit.yaml.

The reason this has not caused a problem yet is because bandit is
still finding a config bandit.yaml in .tox/pep8/etc/bandit/bandit.yaml.
But in a newer version of bandit this will not be the case.

Change-Id: Ie81d052cd57318a987f0ebfe718e1c01bdbffcd3