Updated openstack/openstack Project: openstack/neutron  3aaaabe9f6f0601de49dec750d0784ac9227298d
Tag the alembic migration revisions for Mitaka

Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.  So
now we tag the final revision on every branch when we make a
milestone release.

The database can be upgraded with the command:
  neutron-db-manage upgrade <milestone>
where <milestone> is an alias for all the revisions for a
milestone.

Partial-Bug: #1552935

Change-Id: I44456a8389b761bc9e0f35845b90671daddf3b59