Updated openstack/openstack Project: openstack/os-cloud-config  e83765604348a95a4a9fd2fc045c475ac8ae5dd8
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

Change-Id: I0824d8adcbf409dc024cda9479bf72782b647fb7
Closes-bug: #1489059