Updated openstack/openstack Project: openstack/neutron-fwaas  75690bcf899e960b5dea88baf6d53cbc72949835
Tag the alembic migration revisions for Mitaka

Previously when we had one repo with one alembic branch we would
create a milestone revision on that single branch. Now we have
multiple repos and expand/contract branches for each repo.  So
now we tag the final revision on every branch when we make a
milestone release.

The database can be upgraded with the command:
  neutron-db-manage upgrade <milestone>
where <milestone> is an alias for all the revisions for a
milestone.

Partial-Bug: #1552935

Change-Id: Id8b012dc3f187b5a4b7be73a69910fe35657a0f5