Updated openstack/openstack Project: openstack/ironic  ff43e845228ab36d235487694fef807dbd46ed56
Don't use token for glance & check for some unset vars

There is a regression issue in stable/liberty where using a token for
auth is failing for glance. We don't need to use a token as we already
have our credentials. So stop using a token to upload to glance.

We call various commands to set certain variables. Add some checks to
make sure those variables are set and fail if they are not.

Partial-Bug: #1549095
Change-Id: Ic941d061edf3fe68c18b3eda198b1ed03f4a18fc