Updated openstack/openstack Project: openstack-infra/project-config  7750941926e5a8738254e1e2311acd1814ea812b
Use ubuntu-trusty for various tox tests not using database

Infra is moving all jobs to use ubuntu-trusty instead of bare-trusty.

As part of this move:
Move tox tests to use 'gate-{name}-tox-nodb-{envlist}' instead
of 'gate-{name}-tox-{envlist}'.

This moves the jobs from running on bare-trusty to ubuntu-trusty. Also,
we use the nodb macro to indicate that no MySQL or PostgresQL database
needs to be setup for these jobs.

Change-Id: I65bea77ebae9154d412f027879d9835380313e12