Updated openstack/openstack Project: openstack/diskimage-builder  5b6716cee8bb69c5f309ce19e7e47da1f74992d9
Use fstrim to prep the block device

This cuts the image size down alot, esspecially if there were lots of
small file deletes.

The fstrim utility is in the util-linux package and should be on
most all systems. fstrim also works with XFS, ext4, btrfs, etc
prodiving the kernel is new enough.

A reduction of 25% or more in size is common.

Change-Id: I269b4416be450369616f9b8e030f84c30e329804