Updated openstack/openstack Project: openstack/keystone  49ef16bf503325c16f3df3aa381e490c2ed1bcf0
Add notifications to user/group membership

When a user is added or removed from a group we should emit a notification. The
notification has group as the resource type and the group ID as the
resource ID. The notification also includes the user, user ID, and the
operation that was done ('removed' or 'added').

This way consumers are notified of user and group memberships.

Change-Id: I93ca3a0cb2fe9b93f5370e5871be1f1b30d87f72
Closes-Bug: 1552639