Updated openstack/openstack Project: openstack/nova  fcdfe50c4a487aac18e61d6df0a9da04d9d46d37
Fix processing of libvirt disk.info in non-disk-image cases

In Idfc16f54049aaeab31ac1c1d8d79a129acc9fb87 a change was made
that caused non-disk-image backends to fall over because of an
undefined variable because they skipped processing of the disk.info
file. This adds a check for that case to make sure we don't run
that path in the non-disk-image backend case.

Closes-Bug: #1555287

Change-Id: I02f8a5f0e29816336e500a8fe8dcc9ece15968e9