Updated openstack/openstack Project: openstack/neutron  c220c9edbeea1ce15e1396c14f62e0ca5ea51b10
Queries for DVR-aware floating IP next-hop lookups

Introduce calls to collect host routes for floating IP's in a
DVR-aware way.  When a floating IP is associated through a
distributed router, the next-hop for the floating IP is announced
as the floating IP agent gateway port on the corresponding host.

Change-Id: I654e4d2496e19c011653eedb87c9b6026f801e9f
Implements: blueprint bgp-dynamic-routing
DocImpact: Neutron BGP should have docs created for the feature