Updated openstack/openstack Project: openstack/neutron  692b4ecef71ce13004f3c618a9a404eaac62cc74
Updates external DNS service if IP address changes

Previously the external DNS service was only notified when dns_name
was updated in a port, but it should also be alerted when the IP
address changes. This patch propogates the changes invoked in a port
update to the external DNS service when either the dns_name or the IP
address of a port changes.

Change-Id: I319f7a3265caaee5edaab19695a99d84e8e9a8ab
Closes-Bug: 1554207