Updated openstack/openstack Project: openstack/heat  73fc655d1c58f73671d22ddf5d3817c82be5e235
Refactor validation of arguments to 'repeat' intrinsic function

Separate validation of the structure of the function call itself from
validation of the arguments passed. The former remains in the
initialisation method, while the latter moves to the validate() method.
This is consistent with how other intrinsic functions are instantiated,
and should hopefully lead to less confusion of the kind that resulted in
bug 1553306.

Change-Id: Ie20b79c33bded82db3befd3ca25fa76ebed3d417
Related-Bug: 1553306