Updated openstack/openstack Project: openstack/nova  e00ffa011339a0aeaceb40528788aeb26d8eb69d
tests: Add some basic compute_api tests for attaching volumes

It is really interesting that we never tested most of this logic even
though there is quite a bit of stuff going on. We only have some basic
tests in test_compute.py that do not check for different failure
conditions.

This patch adds some basic unit tests for the basic case of attaching a
volume to a non-shelved instance, and a failure to reserve a volume.

Change-Id: I6d3745dc17cd8534bd90c4d76448434bdc26ae83