Updated openstack/openstack Project: openstack/nova  47a42c4e2fc1b195c0845c427657ca8dfef6452e
xenapi: fix when tar exits early during download

When downloading a streamed chunk from swift through glance gives tar
invalid input, tar will exit, but leave the plugin still trying to write
data to the dead process.

To stop this we can spot when the tar process exits early, and be sure
to stop trying to write more data to the dead process.

Change-Id: Ic2bc89fa6d08db505b044a9498c1bfa5b884a056
Closes-Bug: 1552293