Updated openstack/openstack Project: openstack/nova  1ea09002e4b329633dd41511239eac11dc03f46e
Tweak the resize_confirm_window help text

The resize_confirm_window config option is used for
both cold migration and resize (where resize is really
just cold migration with a different flavor). But
you might not know this unless you looked at the periodic
task that uses the option.

So this change updates the help text slightly to also
note that it's used for automatically confirming
cold migrations.

Change-Id: I4e21a848c4fa483b5975f6b09d6d05c161922277