Updated openstack/openstack Project: openstack/nova  681f584186afbb7e254be5e0d9a6f012a489ce41
Use SensitiveStringField for BlockDeviceMapping.connection_info

bd977f400a1192d5cf7c2b52ef91615c0828813c added the SensitiveStringField
to oslo.versionedobjects. SensitiveStringField is a StringField, which
is what BlockDeviceMapping.connection_info was already using. The difference
is that a SensitiveStringField masks passwords in the 'stringify' method,
which is what's used when __repr__ is called on the object.

Since BDM.connection_info can contain credentials, and the connection_info
dict gets passed around quite a bit in the compute manager and virt drivers,
it has from time to time gotten logged without first masking passwords.

This makes the object handle masking the password so we don't have to do it
explicitly anymore.

There is no version bump on the BlockDeviceMapping object since nothing has
functionally changed.

Change-Id: I66a0b5f6834034e2fcbefc4510e3aa018edec310
Closes-Bug: #1321785