Updated openstack/openstack Project: openstack/nova  748833d3cf43a43112801ec10a9900292fe741a1
plugins/xenserver: Resolve PEP8 issues

Before enabling PEP8 checks for these files, the actual PEP8 issues
need to be resolved. There are a number of "types" of issues that need
to be fixed:

* E112 expected an indented block
* E241 multiple spaces after ','
* E265 block comment should start with '# '
* E302 expected 2 blank lines, found 1
* H102 Apache 2.0 license header not found
* H237 module module is removed in Python 3
* H303 No wildcard (*) import.
* H404 multi line docstring should start without a leading new line
* H231 Python 3.x incompatible 'except x,y:' construct
* F403 from module import *' used; unable to detect undefined names
* F841 local variable 'name' is assigned to but never used
* N344 Use six.iteritems() instead of dict.iteritems().

The fixes for the above are all pretty self-explanatory, but there is
one exception. The 'N344' issues are not ignorable due to how the
"hacking" check is implemented. Rather than ignoring them, skip the
issue altogether by replacing 'iteritems' with 'items'. The difference
in memory usage that this incurs should not be noticable.

Change-Id: Ie26e3daca3b408c9ea43613bcb9996618b9abc69
Partial-Bug: #1302831