Updated openstack/openstack Project: openstack/nova  70499495ced694a2bd1f689ecd3a02435d1465b4
Set personality/injected_files to empty list if not specified

Previously, in the legacy_v2 controller, a "create" would pass an empty
list as a default for the injected files to the compute API. This
behavior has changed, and now, if no personality is specified, the
injected_files parameter ends up as None.

An empty list coming from the controller enabled us to inject files
through the hook mechanism. So, having this behavior changed,
breaks the functionality that was available through this, since we
now aren't able to use that parameter effectively.

Change-Id: Iac02c4ba23381889c36d897451b3e89756a75d4d
Co-Authored-By: Rob Crittenden <rcritten@redhat.com>
Closes-Bug: #1518321