Updated openstack/openstack Project: openstack-dev/devstack  72bd9c81f5d42980958cd100d14df663f40975ce
Ironic: Set DEFAULT_IMAGE_NAME to full disk image for agent

The agent_* drivers in Ironic are not yet capable of deploying partition
images. The code in DevStack was aware of that but it was looking at the
agent_ssh driver specifically.

This patch is fixing this assumption and extending that conditional for
all agent drivers.

Change-Id: I416faa9ef6fba9621cd664d5a0747b1e80e281d1