Updated openstack/openstack Project: openstack/tripleo-heat-templates  31ffe53d75dd81a9ff083224f0ae7d509ae77f96
Upgrades: object storage node upgrade fix

The variables in the heredoc should be escaped because they should
evaluate only when the inner script runs, not when the outer "writer"
script runs.

Python-zaqarclient is installed for os-collect-config to work, as we do
on the other node types.

Swift-proxy is removed from list of services to stop/start, as
swift-proxy isn't supposed to run on the swift storage nodes.

Change-Id: I8426b859d11378ebdc3da94dcc090133dab0c628