Updated openstack/openstack Project: openstack/cinder  e41a9820cf8f343e1444a4a176e75d6c25e50a9a
Show qos_specs_id based on policy

Right now qos_specs_id is only shown to an admin user
when showing a volume type. This patch changes that to
be based on policy to allow for more flexibility. It
also adds unit tests for showing a volume type
with policy permissions for qos_specs_id as well as
extra_specs.

APIImpact

Change-Id: I4e6e99b8992b6941ba247bee90493cc2adba7f0b
Closes-bug: #1512876