Updated openstack/openstack Project: openstack/cinder  b31088f0a6a566eaca2e9cba9c4ecc4a4a974692
NetApp: volume resize using clone fails with QoS

Data ONTAP cannot resize a LUN past the geometry established
when the LUN was created. So to resize past that limit, the
DOT drivers create a new LUN of the needed size and clone
the original LUN into it. This process doesn't always work
if a QoS policy is in place, and the fix is to not pass the
QoS policy to the clone operation.

The fix is trivial, and I improved the surrounding code a
little, but there was no unit test coverage for the method
in question, so this commit also adds full coverage for the
LUN clone method.

Change-Id: I04dcf7ba329a9b94a39580b1115a9dec41231b85
Closes-Bug: #1550478