Updated openstack/openstack Project: openstack/glance_store  851508c2bff70d2857f7d0cdb9eb684f0288450b
Mock swiftclient's functions in tests

swiftclient changed their code so if the chuck size is greater than
zero it will return a generator otherwise it will return the response.
Which would be the object you are looking for.  This causes our tests
to fail since they are always expecting the response instead of a
generator.

Closes-bug: 1556059

Co-Authored-by: avarner <avarner@us.ibm.com>
Change-Id: I62cc037493af87373a75e37e2d9f33b8aedf9889