Updated openstack/openstack Project: openstack/keystone  ecb5613147fc8d6b1fd83b7016554ca05d39cad3
Fix keystone-manage example command path

Whether the user is using tox or has installed keystone globally, the
keystone-manage command should already be in their path. The keystone
repo doesn't have a bin/ directory so trying to call
bin/keystone-manage from the root of the repo will result in an error.
Moreover, other references to keystone-manage in this page do not
specify a path. This patch fixes the instruction and makes it
consistent with the rest of the document.

Change-Id: I93f1aca9bdef0fa3ef4937ef616157d1e1b6fbe9