Updated openstack/openstack Project: openstack/keystone  e4e16cefab34d81c155b4814338a648e0a64b1b9
Fixing mapping schema to allow local user

Mapping to local user was broken since we introduced JSON schema for
mapping. That was because we've never accounted for the "type" attribute
in the "user" object. This patch add it back so that we can map a
remote user to either a local or ephemeral user.

Closes-Bug: 1555403

Change-Id: I7ae69b143e1bdade3e8ea9e2036227c7a38b3b10