Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  630ba7fad2c0c3352ff53549a7f08d78a9541305
Switch use of cStringIO.StringIO to io.BytesIO

This commit switches all our uses of cStringIO.StringIO to io.BytesIO.
The StringIO class is the wrong class to use for this, especially
considering that subunit v2 is binary and not a string. This commit
fixes it so we're using the proper type.

Change-Id: I5d21f7ca2f40cbd5c2659f6cee1570ebe2d4a983