Updated openstack/openstack Project: openstack-infra/system-config  4978adf39e289077290d82273cacfe88434230b5
Logstash 2.0 compat ES output rule

Logstash 2.0 defaults to HTTP elasticsearch output which means that the
elasticsearch output (no _http) does HTTP and a new elasticsearch_java
output exists if you want to continue doing the native api output. We
had been doing HTTP so just need to update the output name. The host
parameter is also deprecated and you must pass an array to the hosts
parameter instead so update that as well.

Note that this switches from using a local ES daemon to talking to the
cluster itself directly because new logstash is able to load balance
over http. This reduces the overhead necessary to have resilient ES
connections.

Note this is not compatbile with Logstash 1.3.3 which is what we are
currently running so this change should only go in as part of an upgrade
to Logstash 2.0 and beyond.

Change-Id: I788ecb936f9fa5a006332ed626f90c33a255d9bf