Updated openstack/openstack Project: openstack/glance  e98bf0df6de798f39f8b8a0c0fe2038f427ed9e4
Test tag against schema to check length

This bug checks the tag against the v2 schema so that if the tag is
too long this is picked up and a 400 in returned along with the schema
message. The schema was used so if it ever changes this will evolve
with it.

Closes-Bug: 1488174
Change-Id: Ibb8250851b1b716d8198ad1a4ca04b7d6a4d40dc