Updated openstack/openstack Project: openstack-infra/puppet-exim  ae24f6250d8211b21c177e5c72fda146822133b8
Always specify driver first

In general the options for routers and transports are not sensitive
to order, except that private options for drivers must succeed the
driver option.  Therefore, it is customary to put the driver option
first.

To ensure this, we special case the driver option in the template
expansion.

Also, for the remaining options, put them in sorted order so that
the config files do not needlessly change on puppet runs.

Change-Id: Id5c21aeaede4b5d780403b00eeac4f2947d36ffe