Updated openstack/openstack Project: openstack/cinder  3a439297c98434c6dfa2834ae8077ab29fc68e6a
Fix up failover_host exceptions to preserve states

UnableToFailOver and InvalidReplicationTarget do not save the state
of replication_status currently. This patch adds host.save() in
order to write the change to the db.

In addition, the manager should honor the current replication state
of the host when InvalidReplicationTarget is raised instead of forcing
it into 'enabled' state.

Change-Id: I01af6d70539e518936c74e425c3fad60f4b94581