Updated openstack/openstack Project: openstack/neutron  01a5d9a3c088e54ae78c068408d419ccc53f8ca8
Fix the context passed to get_subnets in _validate_routes

In the patch for bug #1538767 (https://review.openstack.org/#/c/273278),
get_ports() is passed with context.elevated(), while later get_subnet()
for those ports is still passed with the original context. This could
cause "Subnet could not be found" exception for ports only accessible
in context.elevated() but not in the original context.

This commit replaces context with context.elevated() in get_subnet().

Related-Bug: #1538767
Change-Id: I21e73625e0a625a431bfb22f847c3b6f9671daea