Updated openstack/openstack Project: openstack-infra/system-config  7bb7badc487b8d6fccc24f678d2e9f3780843c12
Add roles for CI users

Without this patch, puppet does not idempotently create the openstackci
and openstackjenkins users. Puppet will create the openstackci and
openstackjenkins users, but won't assign them any kind of membership in
the openstackci and openstackjenkins projects. Then on the second
puppet run, puppet tries to check the users' passwords by issuing an
'openstack token issue' command. Without a role, the users can't
authenticate and receive a 401. Puppet then reports that it 'changed
password' because the password check failed.

The name of the role, 'user', is not significant.

The strange syntax of the keystone_user_role resource is explained in
the keystone module[1].

[1] http://git.openstack.org/cgit/openstack/puppet-keystone/tree/examples/user_project_user_role_composite_namevar.pp

Change-Id: I4fb94722ccafb80cdbefa9500b2124a82ddd57cf