Updated openstack/openstack Project: openstack/cinder  2c302cc444f03a8b0861f18c5a4f33cd43e982bc
Allow api_version_request.matches to accept a string

According to the Cinder devref you should be able to use the following
pattern:

    if req_version.matches("3.1", "3.5"):
        ....stuff....
    elif req_version.matches("3.6", "3.10"):
        ....other stuff....
    elif req_version > api_version_request.APIVersionRequest("3.10"):
        ....more stuff.....

However, the api_version_request.matches() function will not accept a
string, it requires an api_version_request object.

Fix this to accept a string, as well as None for object.

Closes-Bug: 1550337
Change-Id: Ic90f4a13cfad601a181fd3a91684ecbb2b2fba74