Updated openstack/openstack Project: openstack/tripleo-heat-templates  dfbcefa0b49fee4b69016bd32638c4d98022dd8b
Upgrades: quiet yum upgrade on cinder nodes

Yum update on cinder nodes should be quiet, as it is on controllers,
because results of these updates are sent to Heat. I mistakenly left
this out in the first patch because i used one of the standalone node
upgrade scripts as a copy/paste base for the cinder node upgrade script.

Change-Id: Id13190dc4d242317829c7994088183f52d21461d