Updated openstack/openstack Project: openstack/cinder  07885de521e078e3861621815092237f2f2af5d2
Pass new volume size when cloning (blockbridge)

When creating a cloned volume (using the Blockbridge driver), pass the
new volume capacity to the backend. After a successful clone, the
backend extends the new volume to the requested size.

Update unit test to ensure capacity param is supplied to the backend
API.

Change-Id: I5e9527691b4575e355143e61fd8ed0354ac2c3d9
Closes-Bug: #1554750