Updated openstack/openstack Project: openstack/oslo.messaging  cc89fc3174e3c0b00d97dc9157be42083c83804a
Revert "Ensure the json result type is bytes on Python 3"

This reverts commit bd81d09c02c5bc8561ad04de91802a5c1917d9e9.

I understand that the change was supposed to fix something, but instead it broke all tests on Python 3!?

It's wrong to replace blindly json.dumps() with jsonutils.dump_as_bytes(). In oslo messaging, the result is usually used as a value in a dictionary, and then the whole dictionary is passed to a second serializer which also serialize to JSON.

Sorry, I don't understand everything, but at least I see that tests passed on py3 before the change, and started to fail with the change.

Maybe json(utils).dumps() is misused in some places, but in this case, you should write a change which only fix these specific places, not replace all calls to dumps().

Change-Id: Icd54ee8e3f5c976dfd50b4b62c7f51288649e112