Updated openstack/openstack Project: openstack/nova  ac9e5baf83cf89dd4409dafb8439a95cc2459df3
Insure resource tracker updated for deleted instances

Instances left in a deleted state at compute manager startup were
not being explicitly cleaned up in the resoruce tracker. To address
this _update_resource_tracker is now called from solely the
_complete_deletion method. That method is called both by the
_init_instance and _delete_instance code paths, both of which should
be doing cleanup in the resource tracker when the vm is DELETED.

Note that this change does mean that if _delete_instance has an
exception, _update_resource_tracker will not be called. This is
considered okay because we cannot guarantee at which stage the
exception happened and the state of the resources involved. For safety
it is better to err on the side of "in use".

Change-Id: I586c0305ea0e20d98697bc9173ca0f89dc2b228c
Related-Bug: #1227925