Updated openstack/openstack Project: openstack/nova  4022d578ccc09e260ad82d684b7f61a6e291e43d
virt: reserved hugepages on compute host

For some use cases we may need to reserved an amount of pages for
third part components.

This commit adds new option 'reserved_memory_pages' which takes a list
of string format to select on which host NUMA node and from which
pagesize we want to reserve a certain amount of pages.

Change-Id: I9d4c07da3594847917c9dc67e6663717d9ab4ba2
Closes-Bug: #1543149
DocImpact: reserved_memory_pages