Updated openstack/openstack Project: openstack-infra/devstack-gate  15767df7db6e03fe5a2d78cfbb8132bd2ad1e89a
Use dd to create swapfile, ensure swap amount == 8GB

It is not possible to use a swapfile generated by fallocate on a
xfs filesystem which is shipped by default on CentOS >7.
So since [1], we effectively ran jobs on CentOS VMs without any
swap.
Use dd as a way of creating the swapfile that works on both Ubuntu
and CentOS with a slight performance hit and ensure it's mounted
properly so we don't have a silent failure.

Also manage a new specific case, Vexxhost VMs ship with 2GB of swap
which is not in line with the expectation that all VMs should have
8GB. If there is less than 8GB of swap (or zero), we will make up
for the difference.

[1]: https://review.openstack.org/#/c/278557/

Change-Id: Iaf4de66475d747f62bd49dfe7318131cbeb03be8