Updated openstack/openstack Project: openstack/neutron  e1251d90a06cfbe702a9eab6734414512f704655
Stops update_network handling updates it shouldn't

Currently update_network() attempts to update the network table with
any parameters that belong to a network, even if those attributes are
stored in a separate database table. This was causing certain update
commands to fail (such as updating dns_domain). The actual update to
dns_domain is handled by extensions called by the ml2 plugin, but this
function was also trying to process it and causing an error.

This patch ensures the network variable is only updated with
attributes that are actually in the networks table by filtering out
any parameters that don't match fields in the Network class model.

Change-Id: I54a736d91de52dda85c559b9e6103199f914abdf
Closes-bug: 1541560