Updated openstack/openstack Project: openstack/cinder  0568d6008fd36c76fb69eb00da77af88f4989b68
Fixup for Pure drivers cheesecake replication setup

We had a remnant of the past replication design left behind in the setup
methods which was causing the arrays to be setup with bi-directional
replication. With the switch to cheesecake we do not need the secondary
arrays to be replicating to the primary. This change removes the step
which would configure that.

Change-Id: I5a0472bec8ccdf3af1e7ce42163b42289b1810bb
Closes-Bug: #1556211